requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
black==20.8b1
certifi==2020.11.8
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
daphne==3.0.1
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.1.4
django-admin==2.0.1
django-allauth==0.44.0
django-cors-headers==3.6.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-extensions==3.1.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-six==1.0.4
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
Faker==5.0.1
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
Pillow==8.0.1
pkg-resources==0.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==20.0.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
rope==0.18.0
screen==1.0.1
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
stripe==2.55.1
text-unidecode==1.3
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
xlwt==1.3.0
zope.interface==5.2.0
